feat: Convert project to TypeScript
Refactors the project to use TypeScript, including type definitions and compilation. Updates build scripts and CLI usage instructions in README.
